Export test-helper to psammead
